+++ src/com/r4intellij/packages/Function.java
@@ -21,14 +21,15 @@ public class Function implements Serializable {
 
     private static final long serialVersionUID = -61945361973531069L;
 
-    private String funName;
+    private final String funName;
 
     private String description;
     private String funSignature;
     private String shortDesc;
 
 
-    public Function() {
+    public Function(@NotNull String funName) {
+        this.funName = funName;
     }
 
     public Function(@NotNull String funName, @NotNull String funSignature) {